Move block stream logic into `ethrpc`
