+++ crates/autopilot/Cargo.toml
@@ -25,6 +25,7 @@ contracts = { path = "../contracts" }
 database = { path = "../database" }
 derivative = { workspace = true }
 ethcontract = { workspace = true }
+ethrpc = { path = "../ethrpc" }
 futures = { workspace = true }
 gas-estimation = { workspace = true }
 observe = { path = "../observe" }
