+++ crates/driver/src/boundary/liquidity/mod.rs
@@ -5,12 +5,13 @@ use {
         infra::{self, blockchain::Ethereum},
     },
     anyhow::Result,
+    ethrpc::current_block::CurrentBlockStream,
     futures::future,
     itertools::Itertools,
     model::TokenPair,
     shared::{
         baseline_solver::BaseTokens,
-        current_block::{self, CurrentBlockStream},
+        current_block,
         http_client::HttpClientFactory,
         recent_block_cache::{self, CacheConfig},
     },
