+++ crates/ethrpc/Cargo.toml
@@ -11,6 +11,7 @@ path = "src/lib.rs"
 doctest = false
 
 [dependencies]
+anyhow = { workspace = true }
 async-trait = { workspace= true }
 ethereum-types = { workspace = true }
 futures = { workspace = true }
@@ -19,6 +20,7 @@ hex-literal = { workspace = true }
 lazy_static = { workspace = true }
 mockall = { workspace = true }
 observe = { path = "../observe" }
+primitive-types = { workspace = true }
 prometheus = { workspace = true }
 prometheus-metric-storage = { git = "https://github.com/cowprotocol/prometheus-metric-storage", tag = "v0.4.0" }
 reqwest = { workspace = true }
@@ -26,6 +28,7 @@ scopeguard = "1"
 serde = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true, features = [] }
+tokio-stream = { version = "0.1", features = ["sync"] }
 web3 = { workspace = true }
 contracts = { path = "../contracts" }
 ethcontract = { workspace = true }
