+++ crates/ethrpc/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod buffered;
+pub mod current_block;
 pub mod dummy;
 pub mod extensions;
 pub mod http;
