+++ crates/refunder/Cargo.toml
@@ -13,6 +13,7 @@ clap = { workspace = true }
 contracts  = { path = "../contracts" }
 database = { path = "../database" }
 ethcontract = { workspace = true }
+ethrpc = { path = "../ethrpc" }
 futures = {workspace = true}
 gas-estimation = { workspace = true }
 lazy_static = { workspace = true }
