+++ crates/shared/src/event_handling.rs
@@ -1,8 +1,5 @@
 use {
-    crate::{
-        current_block::{BlockNumberHash, BlockRetrieving, RangeInclusive},
-        maintenance::Maintaining,
-    },
+    crate::maintenance::Maintaining,
     anyhow::{Context, Error, Result},
     ethcontract::{
         contract::{AllEventsBuilder, ParseLog},
@@ -11,6 +8,7 @@ use {
         Event as EthcontractEvent,
         EventMetadata,
     },
+    ethrpc::current_block::{BlockNumberHash, BlockRetrieving, RangeInclusive},
     futures::{future, Stream, StreamExt, TryStreamExt},
     std::sync::Arc,
     tokio::sync::Mutex,
@@ -574,12 +572,13 @@ fn track_block_range(range: &str) {
 mod tests {
     use {
         super::*,
-        crate::{
-            current_block::block_number_to_block_number_hash,
-            ethrpc::{create_env_test_transport, Web3},
-        },
         contracts::{gpv2_settlement, GPv2Settlement},
         ethcontract::{BlockNumber, H256},
+        ethrpc::{
+            create_env_test_transport,
+            current_block::block_number_to_block_number_hash,
+            Web3,
+        },
         std::str::FromStr,
     };
