+++ crates/shared/src/price_estimation/http.rs
@@ -416,7 +416,6 @@ mod tests {
     use {
         super::*,
         crate::{
-            current_block::current_block_stream,
             gas_price_estimation::FakeGasPriceEstimator,
             http_solver::{
                 model::{ExecutedAmmModel, ExecutedOrderModel, InteractionData, UpdatedAmmModel},
@@ -445,7 +444,7 @@ mod tests {
         anyhow::anyhow,
         clap::ValueEnum,
         ethcontract::dyns::DynTransport,
-        ethrpc::{http::HttpTransport, Web3},
+        ethrpc::{current_block::current_block_stream, http::HttpTransport, Web3},
         gas_estimation::GasPrice1559,
         maplit::hashmap,
         model::order::OrderKind,
